requirements.txt
httpx>=0.27,<1
python-dotenv>=1.0,<2
rich>=13,<14
